resolve cmt #880
