+++ libs/mobile-ui/src/lib/themes/Colors.ts
@@ -18,6 +18,7 @@ const colors = {
 	tertiaryWeight: isDarkTheme ? '#1E1E1E' : '#E1E1E1',
 	gray48: isDarkTheme ? '#484848' : '#484848',
 	textLink: '#007AFF',
+	borderDim: isDarkTheme ? "#5a5b5c30" : "#5a5b5c30"
 };
 
 export default colors;
